# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "CRM API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/owner/user/list": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "List User",
                "description": "List User",
                "operationId": "UserList",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "no of record you already get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "no of record you want to get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by name, mobile, email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column(eg. id, name, mobile, email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/user/store": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store a User",
                "description": "User",
                "operationId": "UserStore",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Enter First Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User First Name"
                    },
                    {
                        "name": "middle_name",
                        "in": "query",
                        "description": "Enter Middle Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User Middle Name"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Enter Last Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User Last Name"
                    },
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "Enter User Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User User Name"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test@yopmail.com"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Enter Role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Admin"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "password123"
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "description": "Enter Department",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "IT"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/user/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update a User",
                "description": "User",
                "operationId": "UserUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter User ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Enter First Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User First Name"
                    },
                    {
                        "name": "middle_name",
                        "in": "query",
                        "description": "Enter Middle Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User Middle Name"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Enter Last Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User Last Name"
                    },
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "Enter User Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "User User Name"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test@yopmail.com"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Enter Role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Admin"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "password123"
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "description": "Enter Department",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "IT"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/user/view": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "View a User",
                "description": "User",
                "operationId": "UserView",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter User ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/user/change-status": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change User Status",
                "description": "User",
                "operationId": "UserChangeStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter User ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Enter Status (Active or Inactive)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Active",
                                "Inactive"
                            ]
                        },
                        "example": "Active"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/user/destroy": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Destroy a User",
                "description": "User",
                "operationId": "UserDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter User ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/owner/user/drop-down": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Drop Down",
                "description": "User",
                "operationId": "UserDropDown",
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}